pyproject.toml
[project]
name = "llm_deployer"
version = "0.1.0"
description = "A FastAPI application to build and deploy web apps using an LLM."
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "google-generativeai",
    "PyGithub",
    "requests",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"